Move InfoParams into pyrat-bot-api; unify target as Coordinates
SDK and host had separately-defined InfoParams structs that diverged
on one field: SDK used Option<(u8, u8)>, host used Option<Coordinates>.
Now both re-export the shared type from pyrat-bot-api with
Option<Coordinates>. SDK's build_info converts to Vec2 at the
FlatBuffers boundary.
